Add formatting into the kernel code.
In an effort to remove the requirements on VDO for certain
user mode tools, we've decided to move the formatting code for
VDO into the kernel itself.

When the VDO code starts for a given table line, we will
look at the geometry block for the VDO volume. If the block
is all zeroes, the code will attempt to format the device.
If it is not all zeroes, we will skip the formatting code
and assume the device is already formatted.

In either scenario (formatting or not) we will then try to
load the geometry block for the VDO volume. If it succeeds
we will start running the VDO like our previous code did
before. If the load fails, we will return an error message
to the caller.

Signed-off-by: Bruce Johnston <bjohnsto@redhat.com>