dm vdo: add geometry block structure
Add geometry block structure to the vdo structure. This structure
mimics the existing super_block structure.

Both structures are initialized at the start of the vdo lifecycle and
uninitialized at the very end of the vdo lifecycle. This simplifies
where we unitialize structures.

Update read_geometry() to use new structure.

Signed-off-by: Bruce Johnston <bjohnsto@redhat.com>