Update ci.yml to use org as approvers for outside PRs
Outside PRs will never run automatically. Instead, we ask for one of the team to check the PR to make sure the code is not malicious. If its not they can use a special issue created by the CI to allow for the automatic testing. This change switches from using the person on Triage to the entire organization.