dm vdo: add formatting parameters to table line
Add support for formatting parameters via the table line.

The three parameters are the memory required for the indexer, whether
the index is sparse or dense, and how many blocks are to be used per
allocation slab. Input values will be parsed and verified.

Increment the device module version number to indicate a change
to the table line.

Signed-off-by: Bruce Johnston <bjohnsto@redhat.com>